Move sanity check routines to con.c
